Move Swap Modals to folder
